backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pandas==2.2.1
openpyxl==3.1.2
pdfplumber==0.10.3
anthropic==0.20.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.1
